# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Mini Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server Utama"
        }
    ],
    "paths": {
...
        "/api/v1/master/pelanggan": {
            "get": {
                "tags": [
                    "Master Pelanggan"
                ],
                "summary": "Get list of pelanggan with pagination and search",
                "description": "Get list of pelanggan.",
                "operationId": "5568a3978135865a1181e05d3a001991",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by nama or ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Master Pelanggan"
                ],
                "summary": "Create a new pelanggan",
                "operationId": "1e16e65f50d54ad78a0d31c80f3a9771",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "domisili",
                                    "jenis_kelamin"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "domisili": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "example": "PRIA",
                                        "enum": [
                                            "PRIA",
                                            "WANITA"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil menambah pelanggan"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/v1/master/pelanggan/{id}": {
            "get": {
                "tags": [
                    "Master Pelanggan"
                ],
                "summary": "Get a single pelanggan by ID",
                "operationId": "2730d5f77daa083c325342a2b167f564",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of pelanggan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Pelanggan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Master Pelanggan"
                ],
                "summary": "Update an existing pelanggan",
                "operationId": "07fdad0485f5ac01a25f4b76bbb9c23a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of pelanggan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Budi Santoso Updated"
                                    },
                                    "domisili": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "example": "PRIA",
                                        "enum": [
                                            "PRIA",
                                            "WANITA"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil mengupdate pelanggan"
                    },
                    "404": {
                        "description": "Pelanggan not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Master Pelanggan"
                ],
                "summary": "Delete a pelanggan",
                "operationId": "c5945e6e1a16d158b8787403466c1aac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of pelanggan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus pelanggan"
                    },
                    "404": {
                        "description": "Pelanggan not found"
                    }
                }
            }
        },
...
    },
...
}